Split the interaction logic into its own file
